+++ src/services/organization.ts
@@ -1,43 +1,46 @@
 import { window, Uri, TextDocumentContentProvider, EventEmitter, workspace } from "vscode";
-import { CLI_TENANT_CMD } from "../constants";
 import { ServiceBase } from "./service-base";
 import { AppRegTreeDataProvider } from "../data/tree-data-provider";
 import { GraphApiRepository } from "../repositories/graph-api-repository";
-import { execShellCmd } from "../utils/exec-shell-cmd";
 import { GraphResult } from "../types/graph-result";
 import { Organization, User, RoleAssignment } from "@microsoft/microsoft-graph-types";
 import { clearStatusBarMessage } from "../utils/status-bar";
 import { v4 as uuidv4 } from "uuid";
+import { AccountProvider } from "../types/account-provider";
 
 export class OrganizationService extends ServiceBase {
-	// The constructor for the OwnerService class.
-	constructor(graphRepository: GraphApiRepository, treeDataProvider: AppRegTreeDataProvider) {
-		super(graphRepository, treeDataProvider);
-	}
+
+	// The account provider.
+    private accountProvider: AccountProvider;
+
+    // The constructor for the OwnerService class.
+    constructor(graphRepository: GraphApiRepository, treeDataProvider: AppRegTreeDataProvider, accountProvider: AccountProvider) {
+        super(graphRepository, treeDataProvider);
+        this.accountProvider = accountProvider;
+    }
 
 	// Shows the tenant information.
 	async showTenantInformation(): Promise<void> {
 		// Set the status bar message.
 		const status = this.indicateChange("Loading Tenant Information...");
 
-		// Execute the az cli command to get the tenant id
-		await execShellCmd(CLI_TENANT_CMD)
-			.then(async (response) => {
-				await this.showTenantWindow(response, status);
-			})
-			.catch(async (error) => {
-				await this.handleError(error);
-			});
-	}
+        await this.accountProvider.getAccountInformation()
+            .then(async (response) => {
+                await this.showTenantWindow(response.tenantId, status);
+            })
+            .catch(async (error) => {
+                await this.handleError(error);
+            });
+    }
 
-	// Shows the tenant information in a new read-only window.
-	private async showTenantWindow(tenantId: string, status: string | undefined): Promise<void> {
-		// Get the user information
-		const user: GraphResult<User> = await this.graphRepository.getUserInformation();
-		if (user.success !== true || user.value === undefined) {
-			await this.handleError(user.error);
-			return;
-		}
+    // Shows the tenant information in a new read-only window.
+    private async showTenantWindow(tenantId: string, status: string | undefined): Promise<void> {
+        // Get the user information
+        const user: GraphResult<User> = await this.graphRepository.getUserInformation();
+        if (user.success !== true || user.value === undefined) {
+            await this.handleError(user.error);
+            return;
+        }
 
 		// Get the assigned directory roles.
 		const roles: GraphResult<RoleAssignment[]> = await this.graphRepository.getRoleAssignments(user.value.id!);